+++ app/policies/backoffice/offer_policy.rb
@@ -43,7 +43,10 @@ class Backoffice::OfferPolicy < ApplicationPolicy
   end
 
   def permitted_attributes
-    [:name, :description, :webpage, :offer_type, [parameters_as_string: []]]
+    [:name, :description, :webpage, :offer_type,
+     parameters_attributes: [:type, :name, :hint, :min, :max,
+                             :unit, :value_type, :start_price, :step_price, :currency,
+                             :exclusive_min, :exclusive_max, :mode, :values]]
   end
 
   private