+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 
 ### Fixed
 - Displaying empty links in service details (@goreck888)
+- Autocomplete element styles (@martaswiatkowska)
 
 ### Security